+++ pkg/resource/promtail-config/promtail-config.go
@@ -24,9 +24,23 @@ type values struct {
 
 type promtail struct {
 	ExtraArgs         []string                   `yaml:"extraArgs" json:"extraArgs"`
+	ExtraEnv          []promtailExtraEnv         `yaml:"extraEnv" json:"extraEnv"`
 	Config            promtailConfig             `yaml:"config" json:"config"`
 	ExtraVolumes      []promtailExtraVolume      `yaml:"extraVolumes" json:"extraVolumes"`
 	ExtraVolumeMounts []promtailExtraVolumeMount `yaml:"extraVolumeMounts" json:"extraVolumeMounts"`
+}
+
+type promtailExtraEnvValuefrom struct {
+	FieldRef promtailExtraEnvFieldref `yaml:"fieldRef" json:"fieldRef"`
+}
+
+type promtailExtraEnvFieldref struct {
+	FieldPath string `yaml:"fieldPath" json:"fieldPath"`
+}
+
+type promtailExtraEnv struct {
+	Name      string                    `yaml:"name" json:"name"`
+	ValueFrom promtailExtraEnvValuefrom `yaml:"valueFrom" json:"valueFrom"`
 }
 
 type promtailConfigSnippets struct {
@@ -92,6 +106,17 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
 		Promtail: promtail{
 			ExtraArgs: []string{
 				"-log-config-reverse-order",
+				"-config.expand-env=true",
+			},
+			ExtraEnv: []promtailExtraEnv{
+				{
+					Name: "NODENAME",
+					ValueFrom: promtailExtraEnvValuefrom{
+						FieldRef: promtailExtraEnvFieldref{
+							FieldPath: "spec.nodeName",
+						},
+					},
+				},
 			},
 			Config: promtailConfig{
 				Snippets: promtailConfigSnippets{
@@ -116,6 +141,14 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
     target_label: 'systemd_unit'
   - source_labels: ['__journal__hostname']
     target_label: 'hostname'
+- job_name: kubernetes-audit
+  static_configs:
+  - targets:
+    - localhost
+    labels:
+      kind: audit-logs
+      __path__: /var/log/apiserver/*.log
+      nodename: ${NODENAME:-unknown}
 `,
 					ExtraRelabelConfigs: extraRelabelConfigs,
 				},
@@ -133,6 +166,12 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
 						Path: "/var/log/journal/",
 					},
 				},
+				{
+					Name: "apiserver-logs",
+					HostPath: promtailExtraVolumeHostpath{
+						Path: "/var/log/apiserver/",
+					},
+				},
 			},
 			ExtraVolumeMounts: []promtailExtraVolumeMount{
 				{
@@ -143,6 +182,11 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
 				{
 					Name:      "journal-var",
 					MountPath: "/var/log/journal/",
+					ReadOnly:  true,
+				},
+				{
+					Name:      "apiserver-logs",
+					MountPath: "/var/log/apiserver/",
 					ReadOnly:  true,
 				},
 			},