+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed
+### Added
 
+- Audit logs in promtail config.
 - Add condition for PSP installation in helm chart.
 
 ## [0.0.6] - 2023-09-28